requirements.txt
python-dotenv==1.1.1
langchain==0.3.27
langchain-core==0.3.74
langchain-community==0.3.28
langchain-google-genai==2.1.9
yfinance==0.2.66
markdown==3.8.2
requests==2.32.5
PyYAML==6.0.2